Rename tests and docs
