+++ doc/code/gbs/graph.dense.rst
