+++ doc/code/gbs/graph.max_clique.rst
