+++ doc/code/gbs/graph.resize.rst
