+++ doc/code/gbs/graph.sample.rst
