+++ doc/code/gbs/sample.rst
